add Bungee Repo Spigot Repo. So we can Build one .jar #32
