+++ .idea/encodings.xml
@@ -1,7 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="Encoding">
-    <file url="file://$PROJECT_DIR$/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/Bungee/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/Bungee/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/Master/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/Master/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/Spigot/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/Spigot/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/Spigot/src/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/Spigot/src/src/main/resources" charset="UTF-8" />
   </component>
 </project>
