+++ Bungee/src/main/resources/bungee.yml
@@ -1,0 +1,5 @@
+name: JoinMusicBungee
+author: T0biii
+version: ${project.version}
+main: de.t0biii.joinmusic.bungee.JoinMusicBungee
+description: Bugnee Plugin for JoinMusic
