+++ Master/pom.xml
@@ -1,0 +1,54 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>de.t0biii</groupId>
+        <artifactId>JoinMusic</artifactId>
+        <version>1.0-SNAPSHOT</version>
+    </parent>
+    <artifactId>JoinMusic-Master</artifactId>
+    <dependencies>
+        <dependency>
+            <groupId>de.t0biii</groupId>
+            <artifactId>JoinMusic-Spigot</artifactId>
+            <version>5.0</version>
+        </dependency>
+        <dependency>
+            <groupId>de.t0biii</groupId>
+            <artifactId>JoinMusic-Bungee</artifactId>
+            <version>1.0</version>
+        </dependency>
+    </dependencies>
+    <build>
+        <finalName>JoinMusic</finalName>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-shade-plugin</artifactId>
+                <version>3.2.4</version>
+                <executions>
+                    <execution>
+                        <phase>package</phase>
+                        <goals>
+                            <goal>shade</goal>
+                        </goals>
+                    </execution>
+                </executions>
+                <configuration>
+                    <relocations>
+                        <relocation>
+                            <pattern>org.bstats</pattern>
+                            <!-- Replace this with your package! -->
+                            <shadedPattern>de.t0biii.joinmusic.metrics</shadedPattern>
+                        </relocation>
+                    </relocations>
+                    <artifactSet>
+                        <includes>
+                            <include>de.t0biii:JoinMusic-Spigot</include>
+                            <include>de.t0biii:JoinMusic-Bungee</include>
+                        </includes>
+                    </artifactSet>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
+</project>
