--- src/main/resources/plugin.yml
@@ -1,36 +1,0 @@
-name: JoinMusic
-authors:
-  - T0biii
-  - Fruxz
-  - altrisi
-depend:
-  - NoteBlockAPI
-version: ${project.version}
-website: http://dev.bukkit.org/bukkit-plugins/joinmusik
-main: de.t0biii.music.main.Main
-database: false
-commands:
-  JoinMusic:
-    description: /JoinMusic
-    usage: /<command> [reload | stop | enable | disable]
-    aliases:
-      - jm
-      - joinmusic
-      - joinmusik
-permissions:
-  JoinMusic.command.reload:
-    description: Allows reloading configs
-    default: op
-  JoinMusic.update:
-    description: Will get info about updates
-    default: op
-  JoinMusic.command.stop:
-    description: Allows stopping music from playing
-    default: true
-  JoinMusic.command.disableOwn:
-    description: Allows disabling music on join for themself
-    default: true
-  JoinMusic.use:
-    description: Allows the user to hear the music on join and access the command
-    default: true
-api-version: 1.13
